Change planner so it only updates some days
A user might have already chosen Cooks for some days so when inverting a
day in the planner in the previous version all non automatically
generated results would be lost. This commit changes that so that when a
day is iverted only that day's result and the results of all days with one
cook get updated.